+++ scripts/tsd.sh
@@ -1,6 +1,5 @@
 #!/bin/bash -xe
 
-cp ./test/types/index.d.ts ./build/
-cp ./test/types/index.test-d.ts ./build/
+cp ./test/types/* ./build/
 
 tsd
